# skillmeat/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillMeat API",
    "description": "REST API for SkillMeat collection manager",
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/collections/{collection_id}/artifacts": {
      "get": {
        "tags": [
          "collections",
          "collections"
        ],
        "summary": "List artifacts in collection",
        "description": "Retrieve a paginated list of artifacts within a specific collection",
        "operationId": "list_collection_artifacts_api_v1_collections__collection_id__artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (next page)",
              "title": "After"
            },
            "description": "Cursor for pagination (next page)"
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by artifact type (skill, command, agent)",
              "title": "Artifact Type"
            },
            "description": "Filter by artifact type (skill, command, agent)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionArtifactsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts": {
      "get": {
        "tags": [
          "artifacts",
          "artifacts"
        ],
        "summary": "List all artifacts",
        "description": "Retrieve a paginated list of artifacts across all collections",
        "operationId": "list_artifacts_api_v1_artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page (max 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination (next page)",
              "title": "After"
            },
            "description": "Cursor for pagination (next page)"
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by artifact type (skill, command, agent)",
              "title": "Artifact Type"
            },
            "description": "Filter by artifact type (skill, command, agent)"
          },
          {
            "name": "collection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by collection name",
              "title": "Collection"
            },
            "description": "Filter by collection name"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tags (comma-separated)",
              "title": "Tags"
            },
            "description": "Filter by tags (comma-separated)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "artifacts",
          "artifacts"
        ],
        "summary": "Get artifact details",
        "description": "Retrieve detailed information about a specific artifact",
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Collection name (searches all if not specified)",
              "title": "Collection"
            },
            "description": "Collection name (searches all if not specified)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artifact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artifact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}/upstream": {
      "get": {
        "tags": [
          "artifacts",
          "artifacts"
        ],
        "summary": "Check upstream status",
        "description": "Check for updates and upstream status for an artifact",
        "operationId": "check_artifact_upstream_api_v1_artifacts__artifact_id__upstream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Collection name (searches all if not specified)",
              "title": "Collection"
            },
            "description": "Collection name (searches all if not specified)"
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully checked upstream status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactUpstreamResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artifact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items for this page"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "ArtifactListResponse",
        "description": "Paginated response for artifact listings."
      },
...
      "CollectionArtifactsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactSummary"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items for this page"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "CollectionArtifactsResponse",
        "description": "Paginated response for artifacts within a collection.\n\nReturns lightweight artifact summaries for efficient collection browsing."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}